feat: allow multiple actions in each command
This allows compound commands separated with semicolons, e.g.:

    root: RECT #000000 500 500 500 500; CIRCLE #000000 100 100 100

This performs all of these actions and then syncs the page. There's
still some possibility of inconsistent state because we release the
lock in between each action. If the page is flipped after RECT but
before CIRCLE, for instance, we'll notice some artifacts. If the
panes are in any other state (e.g., is active or is inactive), this
is perfectly atomic.

So there are some loose ends, but this is close to supporting
animations.