feat: add TERMINATE command
This is just a dead-simple text comparison. For commands that don't operate on
any buffers but do take arguments... This won't help. Perhaps it would be
better to have TERMINATE operate like a normal command but only allow it to run
on the root buffer. However, this is simpler and works just fine.

Fixes #10.